feat: add API module
