+++ config/routes.rb
@@ -1,6 +1,7 @@
 Rails.application.routes.draw do
-  resources :orders
-  resources :order_items
-  resources :pizzas
-  # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
+  namespace :api do
+    resources :orders
+    resources :order_items
+    resources :pizzas
+  end
 end
